requirements.txt
python-dotenv
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary,pool]==3.2.12
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==22.0.0
whitenoise==6.6.0
django-extensions==3.2.3
openai>=2.15.0
requests==2.31.0

